修改bug，修改ui，测试

+++ ShuchuangClient/MyTimePickerView.m
@@ -127,7 +127,7 @@
 }
 
 - (NSString*)pickerView:(UIPickerView *)pickerView titleForRow:(NSInteger)row forComponent:(NSInteger)component {
-    return [NSString stringWithFormat:@"%02ld", row];
+    return [NSString stringWithFormat:@"%02ld", (long)row];
 }
 
 @end
